Fix Issue #116
This change prevents file paths or names to be truncated in
omrsl_open_shared_library. OMRPORT_SL_UNSUPPORTED is thrown
if a file path or name is found to be truncated. A test case
is added to check boundary conditions for file paths and
names in omrsl_open_shared_library.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>