Change -fpeel-loops to -funroll-loops
The GCC compiler option -fpeel-loops is not supported by Clang, which
prevents us from using Clang as a drop in compiler.  Clang does,
however, support the similiar option -funroll-loops.

-fpeel-loops is a GCC option which implies -frename-registers, uses
profile feedback to peel loops, and turns on complete loop peeling.

-funroll-loops is a GCC option which implies -frerun-cse-after-loop,
-fweb, -frename-registers, and turns on complete loop peeling.

I think -funroll-loops is the more aggressive optimization.  Since we do
not use profiles to direct optimization, I do not think it will make
much of a difference which option we use. On my machine, switching the
option did not result in identical output files.

Signed-off-by: Andrew Young <youngar17@gmail.com>