Include a dummy install-sh to satisfy configure
configure requires a copy of install-sh. This patch provides a dummy
version to satisfy this requirement. OMR is not installable, so
install-sh is actually unused.

Automake provides an implementation of install-sh, which we cannot check
into the project due to copyright/licensing concerns.

run_configure.mk was originally responsible for creating a dummy
install-sh. We want configure to be runnable on it's own, without the
run_configure.mk wrapper.

Signed-off-by: Robert W Young <rwy0717@gmail.com>