Allow optional pre-compare in compare and swap
Allow callers of atomic compare and swap to optionally specify that a
compare be done once before the atomic operation.  This is useful to
avoid bus contention when the operation is likely to fail.

Currently, this optimization is only performed on x86 and s390
processors, as benchmarking has proven this to be a negative on Power
and no benchmarking has been performed on ARM.

Take advantage of this feature in omrthread spinlocks.

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>