Make sure thread abort tests have a thread to abort
Add synchronization at the beginning of the test so that the dependent
thread (the one that will be aborted) is known to have started when
the abort is issued. With this change, it's possible the dependent
thread won't be actually running or sleeping as the test intends,
but the abort should never be missed which should make the test more
reliable. This patch fixes Issue #112.

Issue: #112
Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>