Add new static assert macros
These macros make use of the c++11/14/newer static_assert feature.

Static assert is especially useful in OMR to document and enforce
requirements on the OMR glue.

If static_assert is not available (IE, the standard is too old), the
macros expand to empty statements. This is safe to do because
static_assert cannot have a side effect on the code.

Signed-off-by: Robert Young <rwy0717@gmail.com>