Issue #189: Detect truncated paths
GetModuleFileNameW retrieves the fully qualified path for the file that
contains the specified module.

DWORD WINAPI GetModuleFileName(
  _In_opt_ HMODULE hModule,
  _Out_    LPTSTR  lpFilename,
  _In_     DWORD   nSize
);

If the length of the path exceeds the size that the nSize parameter
specifies, GetModuleFileNameW will succeed and the path is truncated to
nSize characters including the terminating null character. 

Code has been added to properly detect truncated paths.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>