Reduce indirect memory-memory copies on System z
A common scenario in object-oriented languages is to copy a field from
one object to another. In OMR this is translated into indirect loads and
stores. If the indirect load of the field we are loading is not
referenced anywhere but the store we can use a memory-memory copy
instruction to carry out the load-store sequence.

This path is currently implemented for direct loads. This commit adds
support for indirect loads. The MVC instruction is preferred to the L +
ST instruction sequence as it is one instruction less to decode, it has
smaller path length (6 bytes vs 8 bytes), and in the worst case scenario
the local register allocator may choose to HPR upgrade the L + ST
sequence to a LFH + STFH instruction sequence which is 12 bytes of path
length.

In addition to the above points the MVC instruction under perfect
conditions (assuming no unaligned data, interlocks, cache misses, or
multiprocessor effects) will execute in 1 cycle for copies of length <=
7 where as the L + ST will always execute in 2 cycles.

Issue: #205
Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>