Prefer generating XR when evaluating iconst 0
When loading the immediate value 0 into a 32-bit virtual register prefer
to generate an XR instruction rather than an LHI instruction to save 2
bytes of icache at no cost to latency.

To achieve this optimization we must introduce an additional parameter
to the formerly known genLoadConst API. The new boolean parameter will
dictate whether the API is allowed to generate instructions that set a
condition code (XR in our case). There are scenarios in which the a
constant is loaded between a compare instructions and a branch
instruction in which generating an XR instruction to load the constant
value 0 is functionally incorrect. We avoid this problem by forcing the
API consumer to explicitly state whether setting the condition code is
allowed at the call site.

This commit also documents and renames the API to a more meaningful
name.

Issue: #212
Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>