Build and test JITBuilder in the main makefile
Add JITBuilder to the main makefile, building and testing it
as nesscessary.

If the OMR_JIT build option is enabled, JITBuilder is built.
If tests are also enabled, the tests are also built and can be
run via `make test`.

As a consequence, Travis CI now builds and tests JITBuilder.

Signed-off-by: Lukas Lorimer <lukasl@ca.ibm.com>