Implement a peephole to reduce LHI R,0 to XR R,R
This peephole optimization attempts to reduce LHI R,0 instructions to an
XR R,R instruction to save 2 bytes of icache. This transformation is
possible if and only if there exists an instruction B that sets the
condition code between the LHI instruction A and some instruction C that
reads the condition code.

This peephole optimization is disabled by default pending more work to
identify LHI instructions that are possible candidates for runtime
patching.

Issue: #212
Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>