Improve constraint propagation for constants in VP
VP's `constrainCompileTimeLoad()` now creates an appropriate constraint
after it folds a load into an `iconst`, `lconst`, `aconst`, or
`loadaddr` of a class. Contrary to its name, it would previously fold
*without* creating a constraint, defeating further analysis of
operations consuming the now-constant value.

The constraint for a node that has become constant in this way is
created just as though the node were constant all along, except that it
is not necessarily a global constraint. In particular, the node will
have its flags set as well. `TR::ValuePropagation::replaceByConstant()`
now calls into this logic in order to set flags. (It would already have
added the constraint.)

The handler for `lloadi` now calls `constrainCompileTimeLoad()` to
attempt compile-time evaluation.

Finally, the handler for `aload` used to stop immediately without
creating constraints whenever `transformDirectLoad()` succeeded, which
prevented further analysis of both constants and known objects. This
handler now continues as normal if the node has retained its `aload`
opcode, i.e. if it has been updated to use an object-specific symbol
reference. If on the other hand the node's opcode has changed, an
attempt is made to constrain it as a constant, since it may have been
transformed into `aconst NULL`, for example.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>