Check array element type loads in constrainIaload
VP (value propagation) attempts in `constrainIaload()` to constrain the
class that results from an `aloadi` of the array component type from a
known class. Care must be taken to ensure that the known class really
represents an array type, since it might be a non-array type in cases
where VP is looking at a dead path.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>