Provide convenience specialization for std::deque
Use of c++ standard containers during compilation requires the use of a
custom allocator.  If we use the std::* definitions directly, we lose
any convenience of template and constructor default values.  This
introduces quite a lot of boilerplate to any use of these containers.

Furthermore, the CS2 containers currently targeted for replacement,
those being CS2::ArrayOf and CS2::StaticArrayOf, have potentially
different behaviour when using the [] indexing operator.  CS2::ArrayOf
will potentially resize the container to fit the correct size.  While
exactly duplicating that behaviour is unwanted, as it would force any
user of the indexing operator to provide a default initializer (and
violate seperation of concerns), we can at least provide a
bounds-checked version of the operator such that out-of-bounds accesses
throw an exception rather than cause undefined behaviour.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>