Annotate Symbol hierarchy as extensible
Add the `OMR_EXTENSIBLE` annotation to the Symbol hierarchy.
There were also a few extensible class anti-patterns that were
fixed, including adding a `self()` method to a few classes,
and adding `self()->` when accessing member functions.

In this process, a few methods needed to be outlined. If this
causes performance problems, they can be re-inlined.

Signed-off-by: Lukas Lorimer <lukasl@ca.ibm.com>