Expose Scavenger Evacuate boundaries
Implementations of the concurrent scavenger object access barrier will
need to know the actual address range for evacuate space.  For
performance reasons the barriers need to be implemented in multiple
components, and we cannot just rely on existing
isObjectInEvacuateMemory() API. The exact evacuate address range needs
to be publicly available.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>