Add DataType operators and use where possible
Currently, there is TR::DataTypes and TR::DataType. In a given
situation, it is not clear which data type to use.

These changes are meant to make TR::DataType behave the same
way as TR::DataTypes by adding operators, so that TR::DataType
should be used over TR::DataTypes whenever possible.

The only times TR::DataTypes should be used is when the enum
constants are needed, or to create a TR::DataType from a
value stored directly as an int.

Signed-off-by: Lukas Lorimer <lukasl@ca.ibm.com>