Fix overflow errors in options processing
The JIT fails to compile for IA32 with newer versions of GCC (tested on
GCC 6.2.1). The failures occur because we use values greater than
0x80000000 to set option flag bits.  Changing OptionTable::parm2 to be
unsigned fixes most of the problem, however the storeSinkingLastOpt
option attempts to set parm2 to -1.

Because storeSinkingLastOpt is unique in attempting to use parm2 as a
signed value, change parm2 to be unsigned, and use a static_cast to
coerce the field into accepting storeSinkingLastOpt's -1 value.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>