Add `lint` make target
Add an `lint` make target. This target builds and runs the OMR
linter.

The OMR linter checks for common pitfalls, such as not using `self()->`
when accessing methods of extensible classes from within another method
of the class. Enabling the linter as a make target allows the linter
to be more easily run, and can be easily integrated into continuous
integration systems.

Signed-off-by: Lukas Lorimer <lukasl@ca.ibm.com>