Re-implement TR_CallSite book keeping with deques
Re-implement the tracking of TR_CallSite targets (both actual and
removed) with TR::deque instead of CS2::ArrayOf.  Some heavier lifting
was required for this change as the existing implementation relied on an
external index variable, as well as the ability to grow the container by
indexing a non-existant object.

The forced rethinking of how TR_CallSite uses its containers exposed
some additional opportunites to replace hand-rolled implementations with
built-in container capabilities for both inserting in front, as well as
removing objects contained within.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>