Replace some outOfMemory calls with exception
This commit replaces outOfMemory calls with exception throws where an
error code is set. There are still several outOfMemory calls in
the code but they will be cleaned up in a subsequent commit. These
changes are part of the plan to decouple the concepts of error codes and
return codes which are currently used interchangably. By getting rid of
outOfMemory calls, we remove the need to determine the type of error by
checking the error code since we can simply catch the approprate
exception. This commit adds the usage of the following exceptions:

TR::ILGenFailure
TR::CompilationInterrupted
TR::ExcessiveComplexity
TR::CodeCacheError
TR::MaxCallerIndexExceeded
TR::InsufficientlyAggressiveCompilation
TR::GCRPatchFailure

Issue: #380
Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>