Refactor dominators init using deque for the stack
Dominators' initialization does a depth-first walk of successor edges
using a local stack to keep track of which edges still need to be
considered.  Currently the stack is implemented using a CS2::ArrayOf
along with a top-of-stack index variable.

In this change set, we replace the ArrayOf with a TR::deque, and in the
process eliminate the index variable in favour of using facilities like
size() and empty(). Furthermore, to clarify this functions intended
strategy, this change set significantly reduces the redundancy in
tracing.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>