Fix throw to goto transformation
When transforming throw into goto, VP (value propagation) now passes the
exception object to the predicted handler in `ExceptionMeta`, rather
than in an auto. This avoids problems in case the handler reads
`ExceptionMeta` more than once, or in case it overwrites the auto.
Additionally, the transformation can now succeed for every predicted
handler, regardless of the code within the handler.

The two calls to `performTransformation()` are now collapsed into a
single call. Previously the second call could have caused the
transformation to be abandoned when it was already partially done.

When `specializedDesyncCatchBlock` is set, VP now transforms the throw
unconditionally, instead of using `performTransformation()` to ask
permission. Previously an assertion would fail if permission were
denied.

Task 119903
Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>