Allow inhibiting of array splitting in GC_IndexableObjectScanner
Enable inhibition of array splitting in GC_IndexableObjectScanner.

Update MM_Scavenger::shouldRememberObject() to not split arrays when 
scanning for RS overflow and to return as soon as a dependent is
found in new space (if any).

Optimize MM_Scavenger::pruneRememberedSetOverflow() by inhibiting 
unnecessary checks after a dependent object is found in new space.

Signed-off-by: Kim Briggs <briggs@ca.ibm.com>