Use the correct child to optimize primitiveClone
An optimization that transforms calls to Object.clone or calls to
J9VMInternals.primitiveClone to cheaper operations mistakenly assumed
the structure of the trees for them is the same, which leads to a wrong
node being used as the object node during the transformation of
J9VMInternals.primitiveClone. This change will fix the problem and also
remove Object.clone from the surppress-inline list in codegen so that
inliner can inline it when it can.

issue: #405
Signed-off-by: liqunl <liqunl@ca.ibm.com>