Copy object before updating forwarding pointer for CS
While Concurrent Scavenger cycle is in progress, mutator can assess a
new copy of a just forwarded object. Thus, the thread that made the copy
of that object must do it *before* forwarding point is installed. This
is contrary to STW operation, when object copy can be accessed only by
scanning GC threads (when copy cache turns in scan cache) - the object
copying may occur at any point prior to pushing the cache for scanning.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>