Prevent RS pruning for class object remembered during CS cycle.
This is just an extension https://github.com/eclipse/omr/pull/386 to
cover a special case of a class (object).
If during Concurrent Scavenger cycle we add Nursery references to a
class that was pending pruning from RS, we will at the end of cycle,
re-check if this class still has nursery references.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>