Allow default inliner policy to be overridden
Currently the OptimizerManager supports setting a policy for each and
every optimization that is created. The inliner has a policy object
which encapsulates many of the heursitic decisions that need to be made
during inlining, but it obtains this policy from the optimizer - a
global default. This change continues the behaviour of acquiring a
global default policy from the optimizer, but only if the policy has not
been overridden when the inliner was created. This allows the inliner
behaviour to be specialized and changed based on need without needing to
change global default policies.

Task 120255
Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>