Use a deque to keep track of BBInfo objects
The current implementation of the _info container that holds
TR_Dominator's BBInfo objects uses a CS2::StaticArrayOf.  This
implementation replaces the representation with one using a TR::deque.

Additionally, this change redefines TR_Dominators::BBInfo to be a struct
rather than a class with all members being public.  This better
represents how this data type is used and should have no effect (beyond
inconsequential name-mangling changes on platforms that encode the
class-key type in mangled names, like MSVC++)

Unlike _dfNumbers, all indexing operations of _info use the indexing
operator (operator []) and thus, beyond reordering the constructor
parameters, and changing which function gets called to clear the
container, this was essentially a drop-in replacement.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>