Remove node from supportsInduceOSR API
Currently supportsInduceOSR and cannotAttemptOSR accept a node parameter
in their API. This node parameter is used only for logging. The question
of whether OSR is prohibited or supported is a question asked based on
the bytecode index -- nodes do not consider in the equation. Soon we
will want to use these APIs in various IL generators so this commit
removes the node since the IL generator may want to ask the question
before having a node in hand.

Task 112130
Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>