Use a deque for the _dominators container
_dominators is currently represented as a CS2::StaticArrayOf.  This
change set replaces its container implmentation with a TR::deque.

The original implementation makes a semantically consequential use of
the CS2::ArrayOf::ValueAt function (which returns a default value for
indexes that are outside the bounds of the container).  Where that is
used, this change provides a local implementation to maintain
compatibility with its users.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>