Clean up OSR APIs
Currently the OSR APIs conflate testing for OSR support with the notion
of when it is ok to clean-up unused OSR exception edges. You many need
to query for OSR support without running the clean-up if asking
mid-transformation. This change makes it explicit in the API whether you
want clean-up to run or not.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>