Remove inconsistent mechanism to enable or disable fast notify
Currently, we rely upon an environment variable to enable or disable the
fast notify feature. Environment variables can be set by anyone, and all
runtime instances will read the environment variable. So, it is not safe
and reliable to use environment variables. A runtime should use command
line options to enable or disable a feature. Using command line options,
a user can be assured that a feature is enabled or disabled for a
specific instance of the runtime. Also, no third party will be able to
influence the runtime's behaviour. Thus, the current mechanism to enable
or disable the fast notify feature has been removed.

Issue: #478

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>