Using a flag for selecting the notify policy
Currently, we are using the notifyPolicy field in J9ThreadLibrary to
select the notify policy. The field takes 4-8 bytes depending upon
architecture. This task can also be performed by reserving a bit of
J9ThreadLibrary->flags. Thus, the notifyPolicy field in J9ThreadLibrary
has been removed. Instead, a new flag is added to select the notify
policy: signal or broadcast.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>