Disable HCR guard merging when using OSR HCR
In most circumstances the Inliner will merge HCR guards and true virtual
gurds to simplify the trees generated. When using OSR to implement HCR
we want to skip this merging so that the HCR guards are distinct from
the true virtual guards. The HCR guards undergo special processing later
in the compile and merging the guards only complicates this processing
for now real benefit.

As part of this change I am fixing a bug in the removeVirtualGuard code
where it does not check the guard type for equality like
findVirtualGuard does - this is only an issue when there are multiple
guards for a given bytecode which this change makes much more common.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>