Eliminate unused _symsKilledByMustKills
Change the definition of the _symsKilledByMustKills field inside the
TR_UseDefInfo::AuxiliaryData datatype to use a TR::deque instead of a
CS2::ArrayOf for its implementation.  Move the specification of its
required size to its constructor as part of AuxiliaryData's initializer
list instead of using a separate resize operation after it has been
constructed.

Despite the unifying theme, these changes are separated in order to be
able to bisect changes to individual data structures.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>