Resolve cross configure issues for the tools
In PR #383 the tools configuration was significantly improved but
it removed the ability for the tools to be cross configured.  If the
top level configure script can be cross configured the tool script also
requires that ability. This change allows the tools to be cross
configured as long as you specify all of the compile options by hand.

This may not be the ideal end solution but it resolves issues with
runtimes that cross configure OMR.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>