Use multimap in OMRLocalCSE
The HashTable in OMRLocalCSE was a custom multi hash table,
where:

* There were either NUM_BUCKETS (107) or symRefCount buckets.
* The first bucket was reserved for certain nodes
  (eg. that include aiadd or aladd), otherwise they are
  placed based on their reference number or hashing their
  children opcodes and symRefs.
* Buckets were circular linked lists, sorted by ascending
  insert time. The bucket pointed at the newest element,
  so traversing started at hash[bucket]->next, and ended
  at hash[bucket].

These changes do not change the number of buckets, the hashing
algorithm, or the order of nodes within the buckets. Semantically,
they are the same.

Increasing the number of buckets (possibly to UINT_MAX) may change
the semantics slightly, since hash collisions are not currently
explicitly checked.

Signed-off-by: Lukas Lorimer <lukasl@ca.ibm.com>