Fix incorrect ref count in transformObjectCloneCall
The first child of the call to primitiveClone is anchored but it's ref
count is not decremented. This fix will anchor all the children and
call prepareToReplace on the call node before the transformation,
avoiding the need to fix up any single node's ref count. This change
also moves the whole optimization for Object.clone into
J9_PROJECT_SPECIFIC, reducing the number of J9_PROJECT_SPECIFIC code.

issue: #506
Signed-off-by: liqunl <liqunl@ca.ibm.com>