Update JitBuilder struct definition examples
Update JitBuilder struct definition examples to use the new, macro based
API.

Since the LinkedList example overlays a C/C++ struct, it makes sense for
it to use the new API, and should therefore use it. The StructArray
example however, while it does overlay a C/C++ struct, this is not its
primary purpose. Therefore, it seems more appropriate for it to use that
API that lets JitBuilder define the alignment of members.

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>