Remove support for enum types
Properly supporting enum types requires the use of
`std::underlying_type<>` from the C++ 11 type traits library. However,
this feature is not supported in some older compilers supported by OMR
(e.g. GCC 4.4.9).

Removing support for enums is done in a separate commit so that if/when
needed, it can be re-added by simply reverting this commit.

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>