Add ThunkBuilder for libffi-like call building
Provide a new kind of builder object that can call C functions from
an array of argument values. A ThunkBuilder object takes a name, a
return type, a number of parameters, and an array of parameter types.
The compiled code generated by a ThunkBuilder takes two arguments and
returns whatever kind of value was specified by the constructor call.
The first argument is the target call address, and the second argument
is an array of Word sized arguments, which will be passed to the
target method as the types indicated by the ThunkBuilder constructor
call.

Also created jitbuilder/release/src/Thunk.cpp test that, initially,
has only single character and single short parameters with no return
type. To be extended later.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>