Clean up TestDriver and BuilderTest
Some parts of `TestDriver`, such as allocateTestData and
deallocateTestData, were not used by any test. These parts
predate the migration to using gtest, and overlap with the
purpose of gtest.

These parts have been removed. If a future test requires any
of the removed pieces, it can include the required changes to
`TestDriver`, or can extend `TestDriver`.

Some unused functions in BuilderTest tests was also removed.

Signed-off-by: Lukas Lorimer <noryb009@gmail.com>