Add UnsignedOverflow CHK Opcode
Treetop with `UnsignedOverflowCHK` opcode has 3 children which are the
add node and the 2 operands of the add. The 2 operands are hung under
the `UnsignedOverflowCHK` in case that the add node got transformed by
the optimizer and the compiler needs to repeat the operation for
`UnsignedOverflowCHK`. For the same reason, the Opcode of the original
operation node is stored in node extension of the `UnsignedOverflowCHK`.

Issue: #276
Signed-off-by: Shuyu Li <shuyuli@ca.ibm.com>