Rearrange files in JitBuilder .gitignore and Makefile
List files alphabetically so easy to find and change all goals and
lists of files to be one-per-line to make it easier to add/remove
tests/files in the right places. I also consolidated the three
lists of tests into just two lists to make it easier to add/remove
tests: one list is the dependences, one list is the commands to run
all the tests.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>