Fix OperandStackTests and Commit() in VirtualMachineOperandStack
This is why developers should write tests for their own code. Somehow
I managed to reverse the order of writes to the virtual operand stack so
that the stack was written in reverse order. Then I also managed to do the
same thing in writing the OperandStackTests test. Two mistakes == not really
right.

This commit reverses the order in Commit and also in the test.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>