Add more control flow services to BytecodeBuilder and fix existing ones
The set of control flow services accessible to BytecodeBuilder was not complete
and did not include IfCmpEqualZero, IfCmpNotEqualZero, IfCmpLessThan, or
IfCmpGreaterThan. This commit adds those services to BytecodeBuilder.

There was also some scary stuff going on in the control flow services that were
added to BytecodeBuilder to intercept and add the virtual machine state
management. I don't think anything untoward was happening in any code currently
using these services but, at the same time, code evolves. I have created
a set of services in IlBuilder that take a pointer to IlBuilder as a destination
builder object; these services require that the object not be NULL. The
double pointer version still exists and takes care of allocating something
if needed. BytecodeBuilder has versions of these that take BytecodeBuilder
objects. The hiding of the IlBuilder services for BytecodeBuilders should
probably be documented better (future commit).

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>