Restore MM_EnvironmentBase::_exclusiveCount
RAS dump agents are hanging intermittently (especially on builds for 
linux-ppc64-le). This likely relates to removal of the _exclusiveCount
member of MM_EnvironmentBase, which was considered to be a duplicate of
OMR_VMThread::exclusiveCount at the time. But _exclusiveCount is used
exclusively for GC-related concerns and needs to be maintained. 

Fixed by restoring MM_EnvironmentBase::_exclusiveCount and related
usage.

Signed-off-by: Kim Briggs <briggs@ca.ibm.com>